+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- Compendium items.
+- Set for Charge weapon capability.
+### Changed
+- Added Set for Charge capability to compendium weapons.
+
 ## [0.2.18] - 2022-09-26
 ### Added
 - Compendium items.